Added py representation of ntebook
